requirements.txt
mediapipe
opencv-python
numpy
requests
pyserial